bloom-backend/requirements.txt
flask-cors 
Flask
groq
Pillow
flask-limiter
Flask-SQLAlchemy
desktop-use

#pip install -r requirements.txt